Cleaned up projectTree architecture
