+++ src/renderer/utils/useGuid.ts
@@ -1,0 +1,17 @@
+import React from 'react';
+
+function uuidv4() {
+  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, (c) => {
+    /* eslint-disable no-bitwise */
+    const r = (Math.random() * 16) | 0;
+    const v = c === 'x' ? r : (r & 0x3) | 0x8;
+    return v.toString(16);
+    /* eslint-enable no-bitwise */
+  });
+}
+
+export default function useGuid() {
+  // eslint-disable-next-line no-unused-vars
+  const [guid, _] = React.useState(uuidv4());
+  return guid;
+}
